requirements.txt
Flask
azure-cognitiveservices-vision-computervision
msrest
requests
